MS-Lab1

+++ src/interfaces/IGenerator.java
@@ -1,5 +1,6 @@
 package interfaces;
 
 public interface IGenerator {
-   float nextInt();
+   float nextFloat();
+
 }
